requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
appdirs==1.4.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
environs==10.3.0
fake-useragent==1.4.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
playwright==1.42.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.0.1
pyparsing==3.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==2.0.27
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
websockets==10.4
yarl==1.9.4
zipp==3.17.0
